+++ backend/canisters/group_index/impl/src/updates/add_local_group_index_canister.rs
@@ -5,7 +5,7 @@ use canister_tracing_macros::trace;
 use group_index_canister::add_local_group_index_canister::{Response::*, *};
 use ic_cdk::api::management_canister::main::CanisterInstallMode;
 use tracing::info;
-use types::{CanisterId, CanisterWasm, Version};
+use types::{BuildVersion, CanisterId, CanisterWasm};
 use utils::canister::{install, CanisterToInstall};
 
 #[proposal(guard = "caller_is_governance_principal")]
@@ -16,7 +16,7 @@ async fn add_local_group_index_canister(args: Args) -> Response {
             let wasm_version = result.canister_wasm.version;
             match install(CanisterToInstall {
                 canister_id: args.canister_id,
-                current_wasm_version: Version::default(),
+                current_wasm_version: BuildVersion::default(),
                 new_wasm: result.canister_wasm,
                 deposit_cycles_if_needed: true,
                 args: result.init_args,
@@ -64,7 +64,7 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response>
     }
 }
 
-fn commit(canister_id: CanisterId, wasm_version: Version, state: &mut RuntimeState) -> Response {
+fn commit(canister_id: CanisterId, wasm_version: BuildVersion, state: &mut RuntimeState) -> Response {
     if state.data.local_index_map.add_index(canister_id, wasm_version) {
         Success
     } else {