+++ backend/canisters/local_user_index/api/src/updates/join_channel.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
-    ChannelId, CommunityCanisterChannelSummary, CommunityCanisterCommunitySummary, CommunityId, GateCheckFailedReason,
+    ChannelId, CommunityCanisterChannelSummary, CommunityCanisterCommunitySummary, CommunityId, GateCheckFailedReason, Version,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -9,6 +9,8 @@ pub struct Args {
     pub community_id: CommunityId,
     pub channel_id: ChannelId,
     pub invite_code: Option<u64>,
+    #[serde(default)]
+    pub rules_accepted: Option<Version>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -25,5 +27,6 @@ pub enum Response {
     UserSuspended,
     CommunityFrozen,
     NotInvited,
+    RulesNotAccepted,
     InternalError(String),
 }