+++ backend/canisters/user/api/src/updates/send_message_with_transfer_to_channel.rs
@@ -31,6 +31,7 @@ pub enum Response {
     TransferCannotBeZero,
     UserSuspended,
     CommunityFrozen,
+    RulesNotAccepted,
     InternalError(String, CompletedCryptoTransaction),
 }