+++ backend/integration_tests/src/wasms.rs
@@ -2,7 +2,7 @@ use crate::utils::local_bin;
 use lazy_static::lazy_static;
 use std::fs::File;
 use std::io::Read;
-use types::{CanisterWasm, Version};
+use types::{BuildVersion, CanisterWasm};
 
 lazy_static! {
     pub static ref COMMUNITY: CanisterWasm = get_canister_wasm("community");
@@ -30,7 +30,7 @@ fn get_canister_wasm(canister_name: &str) -> CanisterWasm {
     let module = read_file_from_local_bin(&format!("{canister_name}.wasm.gz"));
 
     CanisterWasm {
-        version: Version::min(),
+        version: BuildVersion::min(),
         module,
     }
 }